Add Bootstrap.
We could have downloaded our own version of bootstrap but instead
we have decided to use a Central Distributed Network(CDN) version.

The reason for this is because we want to be kind to our users. If
our users have been to any other bootstrap site that is using the
same version of bootstrap then their browser will use the cached
instead of having to download all the files again.

Its also to our advantage because it will not use up our bandwidth
and help keep web hosting costs down.